Rename pkg name templatelookup
