+++ internal/controller/kyma_controller.go
@@ -43,7 +43,7 @@ import (
 	"github.com/kyma-project/lifecycle-manager/pkg/remote"
 	"github.com/kyma-project/lifecycle-manager/pkg/signature"
 	"github.com/kyma-project/lifecycle-manager/pkg/status"
-	"github.com/kyma-project/lifecycle-manager/pkg/template_lookup"
+	"github.com/kyma-project/lifecycle-manager/pkg/templatelookup"
 	"github.com/kyma-project/lifecycle-manager/pkg/util"
 	"github.com/kyma-project/lifecycle-manager/pkg/watcher"
 )
@@ -567,7 +567,7 @@ func (r *KymaReconciler) updateStatusWithError(ctx context.Context, kyma *v1beta
 }
 
 func (r *KymaReconciler) GenerateModulesFromTemplate(ctx context.Context, kyma *v1beta2.Kyma) (common.Modules, error) {
-	templates := template_lookup.GetRegular(ctx, r, kyma, r.SyncKymaEnabled(kyma))
+	templates := templatelookup.GetRegular(ctx, r, kyma, r.SyncKymaEnabled(kyma))
 	for _, template := range templates {
 		if template.Err != nil {
 			r.enqueueWarningEvent(kyma, moduleReconciliationError, template.Err)
